requirements.txt

neo4j==5.23.0
pydantic==2.8.2
python-dotenv==1.0.1
ollama==0.3.1
langchain==0.2.16
langgraph==0.2.0
tenacity==8.5.0
click==8.1.7
orjson==3.10.7
duckduckgo-search==6.2.11
tavily-python==0.3.6
loguru==0.7.2
